requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
docker==7.0.0
frozenlist==1.4.0
grpcio==1.60.0
grpcio-tools==1.60.0
grpclib==0.4.6
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
multidict==6.0.4
packaging==23.2
protobuf==4.25.1
pydantic==2.5.2
pydantic_core==2.14.5
requests==2.31.0
sentry-sdk==1.38.0
setuptools==69.0.2
typing_extensions==4.8.0
urllib3==2.1.0
wheel==0.42.0
yarl==1.9.4
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
docker==7.0.0
frozenlist==1.4.0
grpcio==1.60.0
grpcio-tools==1.60.0
grpclib==0.4.6
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
multidict==6.0.4
packaging==23.2
protobuf==4.25.1
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
requests==2.31.0
sentry-sdk==1.38.0
setuptools==69.0.2
typing_extensions==4.8.0
urllib3==2.1.0
wheel==0.42.0
yarl==1.9.4
